stats/storage: allow using different storage for analytics
We've noticed that managing the stats buckets is a costly operation in Redis.
In particular, we've seen that the sunion command used in
`BucketStorage#buckets_content_with_values` usually blocks a Redis server for
100-300ms. Also deleting a single bucket usually takes around 10ms.
Moving these operations to a dedicated server will reduce the load of the main
Redis servers.