extensions/resque: memoize all the resque hook methods
These methods, as explained in the comments of the class, are expensive
and need to be memoized. The problem was that we memoized only 5 of the
8 hook methods wasting CPU cycles.

In a simple test, one of the methods that was not memoized, was
consuming like 25% of the CPU time, which represents a small % of
the total time as most of it's spent on Redis, but still worth optimizing.