user: delete unused Redis key helpers
There were 2 helper methods that deal with Redis keys to store users.
One of those helpers generates keys to store the user attrs in a Redis hash.
The other generates a different key for each of the user attrs. The
second strategy is never used, so we can delete those helper methods.