transactor: use correct validators if oauth and didn't lookup a token
We were assigning the incorrect validators to the call when both app
and user ids were being specified. Since those validators vary
subtly, this caused a call to OAuth with both being specified to
fail in one of the validators that change between OAuth and
non-OAuth modes, which is the app_key validator.

In that case, if an app_key exists and both app_id and user_id are
specified, the fact that no app_key is specified is ok in OAuth, but
it's not in non-OAuth. Since the validators run are the non-OAuth,
the call fails the validation unless the correct app_key is
specified.