spec/api/internal: DELETE verbs cannot be used with bodies in rack-test
rack-test switched to not allowing bodies in DELETE requests, so we
can no longer test DELETE requests with bodies. In fact, if we
specify rspec_api_documentation's "raw_post" in DELETE requests,
just as with GET requests, we will not get a body but their contents
added to the URL so that we'd end up with a parameter named
'{"argument":"value"}' with a nil value (in our case, where we use
JSON).

The fix is to just send the parameters as is, which will be
extracted from the URI's query string. This means that the parsing
of bodies of DELETE requests (and GETs) cannot be tested.

Officially there is no prohibition to use bodies in DELETE requests,
although it is usually understood that the URI should provide all
necessary data.

The change in rack-test also means that a default Content-Type
header is not added to DELETE requests (but you can still specify
it).