test/integration/access_token: test array of tokens and adapt to rack-test
We cannot use an empty array as a parameter because it is replaced
with [""] by Rack and newer versions of rack-test.

The empty array was always represented by rack as an array of a
single empty string, [""]. The newer versions of rack-test are now
conforming to what rack does. Since the test now exercises
a different path in this case, we have added a specific test for it,
in which we should get an invalid token format exception.