storage: overhaul interface for Redis client configuration and options
Hopefully this is a bit clearer for users and will now accept Redis
URLs.

The fact that our configuration system is less than ideal and the
options and defaults need to be changed depending on circumstances
makes it hard to deal with.

The helper "config_with" should now generate a correct parameter for
calling Redis.new. The strategy is that we always try to generate
a "url" option. The interface has been extended to be flexible, but
most of the time we will use it with a configuration object and some
options we'd like to ensure are always set (if any).

[eguzki: integrated a fix for host interpreted as path]