util: compute number of processors using Etc package
This uses the scheduler's CPU affinity set or the system's online
processor availability to determine the amount of CPUs.

Note that this changes semantics a bit so that instead of returning
the total CPUs in a system we now return the CPUs this task can be
run on.

This is still not perfect for our usage since we are interested in
the amount of parallel execution units we have available, and some
systems working with cgroups, notably Kubernetes, will show any
number of available CPUs but still not allow us to run on all of
them concurrently. We could even be way more restricted than that,
so a new avenue to work on this in the future would be to determine
rights to processors through the cgroups API.

[alex: changed to use Etc.nprocessors, commit description]