backend: extract Bundler setup code into its own file
This is useful so we don't have to load the whole backend code to
set the environment up (ie. for small or self-contained tasks or
cases in which we don't want backend.rb dependencies, we'd just
require the bundler file and the relevant code).

Further down the road we can identify places where unnecessary loads
are happening and replace the backend.rb require with bundler.rb.