Publish default backend_service port on Makefile 'dev' rule.
In order to be able to perform local development testing more easily we expose and publish the default port (3000) of the 3scale_backend listener
when executing the Makefile 'dev' rule.

The variable PORT has been added to allow the client of the Makefile to remap the local backend_service port to a desired one. If the variable is not defined then the value is the default port of the
backend_service service (3000).

Also, the README.md documentation has been updated accordingly.