Makefile: Add dev-service and dev-service-clean Makefile rules
This rules have been added to be able to start a container with the 3scale_backend service running.

The rule dev-service-clean removes the container.
Also, the dev-clean-image rule has been updated to remove the associated
dev-service container too.
Due to now we may have the dev-service and dev container running we have improved
the running container detection logic in the dev-service and dev rules to only match
the container name (whole name and not matching prefixes)

Finally, we now allow the user to control the command to be executed in the docker container
when executing the 'dev' Makefile rule. If the command is not specified then
/bin/bash is executed by default in the container when calling the dev rule.