Rakefile: Refactor RSpec tasks to individual components
We create different Rake tasks to execute individual sections of our RSpec tests
as we do with Test::Unit tests

When the 'spec' tasks are executed via Rake and some specific test or
tests from an inner subdirectory of spec are executed the spec/spec_helper
is not required by the tests code so some tests have problems not locating
the classes.
This problem has been identified when trying to execute the Rake
spec:api task (that only executes the tests located in spec/api). We require
the spec/spec_helper.