Rakefile: set the environment variable to RACK_ENV if environment is testable
We set RACK_ENV environment variable to "test" if
the environment is testable (currently, this is: if RACK_ENV is not
production neither preview).

This has been done because the test
spec/acceptance/api/internal/utilization_api_spec.rb fails if this variable
is not set due to that test uses a class that includes a new method to itself
if the environment variable RACK_ENV is set to test. Specifically, the
file where this is done is located at
lib/3scale/backend/transactor/notify_batcher.rb