Rakefile: introduce the changelog:link_prs task
This task looks for unlinked issue or PR references in the changelog
and links them to GitHub.

You can use "issue #15" or "Issue #15" to refer to an issue and
force the task to use a link to that issue, or "PR #15" or just
"#15" to link to a pull request.

Valid references must be preceded by a space or an opening
parenthesis. There is room to improve the current regular expression
but it's probably not worth it, as it works well for all practical
cases.