failed_jobs_scheduler: Use Redis Resque instead of Storage for distributed lock
Currently the Redis Distributed Lock is set on the Redis Storage
when the failed jobs scheduler is executed. However, this not allows
to have multiple failed jobs schedulers in execution at the
same time pointing to different Resque servers. Moving the distributed
lock to the Resque server itself solves this problem.