service: Fix set a Service as active when no entry in the DB is found
This is a bugfix commit. Currently when a Service existed in the
database but had no 'state' entry then was marked as :suspended.
This breaks compatibility with previously existing Services.
With this commit existing Services in the database that have no
'state' entry in the database are marked as :active to ensure
compatibility.