usage_report: do not take into account status when calculating remaining quota
Taking into account the status is incorrect. Suppose that we have
several "usage_reports" for the same "status". If the status is
unauthorized, all of them with return a remaining of 0. This is a
problem because we cannot properly compare them to figure out which one
of them is causing the "unauthorized" status. We could have a report
that goes over limits and another that does not.