.circleci/config: clear rbenv cache
The cache for rbenv versions depends on the .ruby-version file. However,
there was a build where the version was 2.4.1 and the image still
contained old Rubies. We dropped support for those Ruby versions and now
the build fails because of that.

The recommended way according to the CircleCI docs to solve this kind of
issues is to use a version prefix in the cache keys. There isn't a way
to clear the cache.