background_job: stop using a class var
This causes problems when using the async mode because all the fibers
share that @args class var. Luckily, with the current code, the only
attribute affected is "enqueue_time", which is not
accurate anyway (see https://github.com/3scale/apisonator/issues/18).

Even without taking the async mode into consideration using a class var
is counter-intuitive. The arguments of a particular job (service ID, app
ID, etc.) should belong to an instance, they shouldn't be stored in a class var.