test/unit/transactor/notify_job: stop sending old timestamps
Notify jobs with an old timestamp raise `TransactionTimestampTooOld`.
The current tests are not failing because they either mock the call that
ends up raising that error or because they test other errors that are
checked before the timestamp.
Let's change this to stop using old timestamps. It will be less
confusing.