validators: add OIDC_VALIDATORS for oauth services calling auth*.xml
Currently OIDC services need to call oauth_authrep.xml instead of
authrep.xml because System is using OIDC applications' client_secrets as
normal application keys.

Those get checked by auth*.xml if existing, so you can't have OIDC
using auth*.xml endpoints and need instead to use the oauth_auth*.xml ones.

This is both a bad decision and an unnecessary consideration from the
POV of the client. The next changes will detect that a service is
configured as OAuth even if not calling oauth_* endpoints, and apply
validators that only check for app_key's if an app_key parameters is
given.