Update puma
Disables `prune_bundler`

It's causing problems on our test suite, and we don't need this really.

prune_bundler is intended to read changes in the Gemfile between phased
restarts, but it's only enabled on development and test, where we don't
need phased restarting.

In production, besides, we use openshift which doesn't
use phased restarts